Fix CLion inspections and Clang-Tidy warnings

+++ src/Simulation.h
@@ -5,4 +5,4 @@
 #include "Vis.h"
 #include <functional>
 
-void simulate(State& state, int ticks, int microticks, Vis *vis, const std::function<Move(const State&, const RobotState&, int)>& getMove);
+void simulate(State& state, int ticks, int microticks, Vis *vis, const std::function<Move(const State&, const RobotState&, size_t)>& getMove);
